Checks module exists before trying to create foreign key
If the module isn't installed than it won't create the foreign key, but it
will still create the column so that the form and model will still work. The
key will need to be created manually if they are not already installed.

+++ protected/modules/OphInGenetictest/migrations/m161219_143706_withdrawal_source.php
@@ -4,14 +4,18 @@ class m161219_143706_withdrawal_source extends CDbMigration
 {
     public function up()
     {
-        //@todo what if related table doesn't exist yet? Annoying modules and migration order.
         $this->addColumn('et_ophingenetictest_test', 'withdrawal_source_id', 'int(10) unsigned');
-        $this->addForeignKey('et_ophingenetictest_test_withdrawal_source_id', 'et_ophingenetictest_test', 'withdrawal_source_id', 'et_ophindnaextraction_dnatests', 'id');
+        if (Yii::app()->modules['OphInDnaextraction']) {
+            $this->addForeignKey('et_ophingenetictest_test_withdrawal_source_id', 'et_ophingenetictest_test', 'withdrawal_source_id', 'et_ophindnaextraction_dnatests', 'id');
+
+        }
     }
 
     public function down()
     {
         $this->dropForeignKey('et_ophingenetictest_test_withdrawal_source_id', 'et_ophingenetictest_test');
-        $this->dropColumn('et_ophingenetictest_test', 'withdrawal_source_id');
+        if (Yii::app()->modules['OphInDnaextraction']) {
+            $this->dropColumn('et_ophingenetictest_test', 'withdrawal_source_id');
+        }
     }
 }
